backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
openai-whisper==20231117
googletrans==4.0.0rc1
python-decouple==3.8